+++ catalyst/contrib/datasets/__init__.py
@@ -1,4 +1,3 @@
 # flake8: noqa
 
 from catalyst.contrib.datasets.mnist import MNIST
-from catalyst.contrib.datasets.metric_learning import MnistQGDataset
