+++ catalyst/contrib/datasets/metric_learning.py
@@ -5,7 +5,6 @@ import torch
 from torch.utils.data import Dataset
 
 from catalyst.contrib.datasets import MNIST
-from catalyst.data.dataset import QueryGalleryDataset
 
 
 class MetricLearningTrainDataset(Dataset, ABC):
@@ -39,6 +38,54 @@ class MnistMLDataset(MetricLearningTrainDataset, MNIST):
             labels of digits
         """
         return self.targets.tolist()
+
+
+class QueryGalleryDataset(Dataset, ABC):
+    """
+    QueryGallleryDataset for CMCScoreCallback
+    """
+
+    @abstractmethod
+    def __getitem__(self, item) -> Dict[str, torch.Tensor]:
+        """
+        Dataset for query/gallery split should
+        return dict with `feature`, `targets` and
+        `is_query` key. Value by key `is_query` should
+        be boolean and indicate whether current object
+        is in query or in gallery.
+
+        Raises:
+            NotImplementedError: You should implement it  # noqa: DAR402
+        """
+        raise NotImplementedError()
+
+    @property
+    @abstractmethod
+    def query_size(self) -> int:
+        """
+        Query/Gallery dataset should have property
+        query size.
+
+        Raises:
+            NotImplementedError: You should implement it  # noqa: DAR402
+        Returns:
+            query size
+        """
+        raise NotImplementedError()
+
+    @property
+    @abstractmethod
+    def gallery_size(self) -> int:
+        """
+        Query/Gallery dataset should have property
+        gallery size.
+
+        Raises:
+            NotImplementedError: You should implement it  # noqa: DAR402
+        Returns:
+            gallery size
+        """
+        raise NotImplementedError()
 
 
 class MnistQGDataset(QueryGalleryDataset):
