+++ catalyst/data/dataset.py
@@ -1,10 +1,8 @@
 from typing import Any, Callable, Dict, List, Union
-from abc import ABC, abstractmethod
 from pathlib import Path
 
 import numpy as np
 
-import torch
 from torch.utils.data import Dataset, Sampler
 
 from catalyst.utils import merge_dicts
@@ -228,59 +226,10 @@ class DatasetFromSampler(Dataset):
         return len(self.sampler)
 
 
-class QueryGalleryDataset(Dataset, ABC):
-    """
-    QueryGallleryDataset for CMCScoreCallback
-    """
-
-    @abstractmethod
-    def __getitem__(self, item) -> Dict[str, torch.Tensor]:
-        """
-        Dataset for query/gallery split should
-        return dict with `feature`, `targets` and
-        `is_query` key. Value by key `is_query` should
-        be boolean and indicate whether current object
-        is in query or in gallery.
-
-        Raises:
-            NotImplementedError: You should implement it  # noqa: DAR402
-        """
-        raise NotImplementedError()
-
-    @property
-    @abstractmethod
-    def query_size(self) -> int:
-        """
-        Query/Gallery dataset should have property
-        query size.
-
-        Raises:
-            NotImplementedError: You should implement it  # noqa: DAR402
-        Returns:
-            query size
-        """
-        raise NotImplementedError()
-
-    @property
-    @abstractmethod
-    def gallery_size(self) -> int:
-        """
-        Query/Gallery dataset should have property
-        gallery size.
-
-        Raises:
-            NotImplementedError: You should implement it  # noqa: DAR402
-        Returns:
-            gallery size
-        """
-        raise NotImplementedError()
-
-
 __all__ = [
     "ListDataset",
     "MergeDataset",
     "NumpyDataset",
     "PathsDataset",
     "DatasetFromSampler",
-    "QueryGalleryDataset",
 ]
